+++ src/chart/types/bubbleData.ts
@@ -1,0 +1,12 @@
+export interface BubbleData {
+  name: string;
+  value: number;
+  color: string;
+  icon?: string;
+}
+
+interface LanguageMapping {
+  color: string;
+  icon: string;
+}
+export type LanguageMappings = Record<string, LanguageMapping>;
